Move idle users fetching logic to a helper function
- Added a new helper function `fetchIdleUsers`
- Used `fetchIdleUsers` helper function inside `availability-panel` page
